Revert "6 спринт на проверку"
