+++ src/com/yandex/app/service/HistoryManager.java
@@ -7,5 +7,4 @@ import java.util.List;
 public interface HistoryManager {
     void add(Task task);
     List<Task> getHistory();
-    void remove(int id);
 }
