+++ src/com/yandex/app/service/InMemoryHistoryManager.java
@@ -8,75 +8,30 @@ import java.util.List;
 import java.util.Map;
 
 public class InMemoryHistoryManager implements HistoryManager {
-
-    private Node head;
-    private Node tail;
-
-    private Map<Integer, Node> taskMap = new HashMap<>();
-
-    private static class Node {
-        Task task;
-        Node prev;
-        Node next;
-
-        public Node(Task task) {
-            this.task = task;
-        }
-    }
-
-    private void linkLast(Task task) {
-        Node node = new Node(task);
-        if (head == null) {
-            head = node;
-            tail = node;
-        } else {
-            tail.next = node;
-            node.prev = tail;
-            tail = node;
-        }
-    }
-
-    private void removeNode(Node node) {
-        if (node.prev != null) {
-            node.prev.next = node.next;
-        } else {
-            head = node.next;
-        }
-        if (node.next != null) {
-            node.next.prev = node.prev;
-        } else {
-            tail = node.prev;
-        }
-        taskMap.remove(node.task.getId());
-    }
+    private List<Task> history = new ArrayList<>();
+    private Map<Integer, Task> historyMap = new HashMap<>();
 
     @Override
     public void add(Task task) {
         if (task == null) {
             return;
         }
-        if (taskMap.containsKey(task.getId())) {
-            removeNode(taskMap.get(task.getId()));
-        }
-        linkLast(task);
-        taskMap.put(task.getId(), tail);
-    }
 
-    @Override
-    public void remove(int id) {
-        if (taskMap.containsKey(id)) {
-            removeNode(taskMap.get(id));
+        if (historyMap.containsKey(task.getId())) {
+            history.remove(task);
+        }
+
+        history.add(task);
+        historyMap.put(task.getId(), task);
+
+        if (history.size() > 10) {
+            Task removedTask = history.remove(0);
+            historyMap.remove(removedTask.getId());
         }
     }
 
     @Override
     public List<Task> getHistory() {
-        List<Task> history = new ArrayList<>();
-        Node node = head;
-        while (node != null) {
-            history.add(node.task);
-            node = node.next;
-        }
-        return history;
+        return new ArrayList<>(history);
     }
 }
